move keycloak init from app.js to a separate file
