+++ demo/mobile/src/app/app.js
@@ -7,9 +7,8 @@ var logger = require('@raincatcher/logger');
 
 // Create INFO logger
 logger.setLogger(new logger.ClientLogger(2));
-var Keycloak = require('keycloak-js');
 
-var module = angular.module('wfm-mobile', [
+angular.module('wfm-mobile', [
   require('angular-ui-router'),
   require('angular-material'),
   require('./services'),
@@ -28,62 +27,7 @@ var module = angular.module('wfm-mobile', [
   require('@raincatcher/vehicle-inspection')
 ]);
 
-// keycloak init config
-var initConfig = {onLoad: 'login-required'};
-
-// the keycloak json config
-var keycloakJS = Keycloak({
-  "realm": "raincatcher",
-  "url": "http://localhost:8080/auth",
-  "ssl-required": "external",
-  "clientId": "raincatcher-mobile",
-  "public-client": true,
-  "use-resource-role-mappings": true
-});
-
-var auth = {};
-
-angular.element(document).ready(function() {
-  // initialise the Keycloak JS Adapter
-  keycloakJS.init(initConfig).success(function() {
-    auth = keycloakJS;
-    auth.name = "keycloak";
-    console.log("Keycloak Ininitalisation Success");
-    // make auth/keycloak JS adapter available to controllers & services in the app
-    module.factory('Auth', function() {
-      return auth;
-    });
-    // angular should be started after Keycloak has initialized otherwise Angular will cause issues with URL Rewrites
-    angular.bootstrap(document, ["wfm-mobile"]);
-  }).error(function(err) {
-    console.error("Error Initialising Keycloak JS", err);
-  });
-});
-
-module.factory('authInterceptor', function($q, Auth) {
-  return {
-    request: function(config) {
-      var deferred = $q.defer();
-      if (Auth.token) {
-        Auth.updateToken(5).success(function() {
-          config.headers = config.headers || {};
-          config.headers.Authorization = 'Bearer ' + Auth.token;
-
-          deferred.resolve(config);
-        }).error(function() {
-          deferred.reject('Failed to refresh token');
-        });
-      }
-      return deferred.promise;
-    }
-  };
-});
-
-module.config(function($httpProvider) {
-  $httpProvider.interceptors.push('errorInterceptor');
-  $httpProvider.interceptors.push('authInterceptor');
-});
-
+require('./keycloak');
 require('./initialisation');
